backend/api-compras/requirements.txt
boto3>=1.26.0
botocore>=1.29.0
PyJWT>=2.6.0
python-dateutil>=2.8.0
